@@ -25,24 +25,24 @@
     "url": "https://github.com/ioBroker/ioBroker.hm-rega"
   },
   "dependencies": {
-    "@iobroker/adapter-core": "^2.6.0",
+    "@iobroker/adapter-core": "^3.0.3",
     "iconv-lite": "^0.6.3",
     "request": "^2.88.2",
-    "xml2js": "^0.6.0"
+    "xml2js": "^0.6.2"
   },
   "devDependencies": {
-    "@alcalzone/release-script": "^3.5.0",
-    "@alcalzone/release-script-plugin-iobroker": "^3.5.0",
-    "@alcalzone/release-script-plugin-license": "^3.5.0",
+    "@alcalzone/release-script": "^3.6.0",
+    "@alcalzone/release-script-plugin-iobroker": "^3.6.0",
+    "@alcalzone/release-script-plugin-license": "^3.5.9",
     "@iobroker/testing": "^4.1.0",
-    "axios": "^1.2.0",
-    "chai": "^4.3.4",
-    "eslint": "^8.20.0",
-    "eslint-config-prettier": "^8.4.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "axios": "^1.4.0",
+    "chai": "^4.3.7",
+    "eslint": "^8.46.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
     "gulp": "^4.0.2",
-    "mocha": "^10.1.0",
-    "prettier": "^2.6.2"
+    "mocha": "^10.2.0",
+    "prettier": "^3.0.1"
   },
   "bugs": {
     "url": "https://github.com/ioBroker/ioBroker.hm-rega/issues"
@@ -51,7 +51,11 @@
   "scripts": {
     "test:package": "mocha test/packageFiles --exit",
     "test:integration": "mocha test/integrationAdapter --exit",
-    "release": "release-script"
+    "release": "release-script",
+    "release-patch": "release-script patch --yes",
+    "release-minor": "release-script minor --yes",
+    "release-major": "release-script major --yes",
+    "update-packages": "ncu --upgrade"
   },
   "files": [
     "main.js",
