@@ -73,7 +73,7 @@ class Rega {
     } // endInit
 
     /**
-     * Check time difference between Rega and System
+     * Check the time difference between Rega and System
      *
      * @return {Promise<number>}
      */
@@ -94,7 +94,7 @@ class Rega {
     }
 
     /**
-     *  run a script file from file system by given path
+     *  run a script file from a file system by given path
      *
      * @param {string} script path to script
      * @returns {Promise<string>}
@@ -146,7 +146,7 @@ class Rega {
 
         this.logger.debug(`--> ${this.pendingRequests[0].script.slice(0, 80).replace(/\n/g, ' ')}`);
         let auth = `${this.username}:${this.pass}`;
-        auth = `Basic ${new Buffer(auth).toString('base64')}`;
+        auth = `Basic ${Buffer.from(auth).toString('base64')}`;
         const timeout = 90 * 1000;
 
         const post_options = {
@@ -179,7 +179,7 @@ class Rega {
                 parser.parseString(xml, (err, result) => {
                     const task = this.pendingRequests.shift();
 
-                    // if connection lost during the process pending requests (task) will be empty
+                    // if connection lost during the process, pending requests (task) will be empty
                     if (task && typeof task.callback === 'function') {
                         if (result && result.xml) {
                             task.callback(stdout, result.xml);
@@ -202,7 +202,7 @@ class Rega {
                         }
                     }
 
-                    // try next task
+                    // try the next task
                     if (this.pendingRequests.length) {
                         setTimeout(() => this.script(), 50);
                     }
